Linux 2.3.1pre3
As to 2.3.x, we're beginning with a long overdue waitqueue cleanup, which
means that a lot of small details need to get fixed in a variety of files.
A working pre-patch of this is to be found as pre-patch-2.3.1-3, but not
all drivers have been fixed - and help is appreciated (even drivers that
_have_ been fixed have not necessarily actually been tested due to lack of
hardware).

	Linus

+++ kernel/sched.c
@@ -907,7 +907,7 @@ void __wake_up(wait_queue_head_t *q, unsigned int mode)
 				goto out;
 			}
 #if WAITQUEUE_DEBUG
-			curr->__waker = (int)__builtin_return_address(0);
+			curr->__waker = (long)__builtin_return_address(0);
 #endif
 			wake_up_process(p);
 		}
